Separate phase part from phase
